so many changes

+++ initialize_folder.py
@@ -5,6 +5,6 @@ import glob
 def initialize_folder():    
     files = glob.glob('/users/josh.flori/desktop/demo/*')
     for f in files:
-        if 'template.jpg' not in f and  'music.wav' not in f:
+        if 'template.jpg' not in f and  'music.wav' not in f and 'padding' not in f:
             os.remove(f)
